fix: Support long version values in element-template-generator
- Changed version field from int/Integer to long/Long throughout the codebase
- Updated ElementTemplate annotation to accept long version values
- Updated ElementTemplate record, builder, and related classes
- Updated PropertyGroup and CommonProperties to use Long
- Updated GeneratorConfiguration to use Long for templateVersion
- Added logging when skipping unparseable template files in maven plugin
- Added test coverage for long version values
- Fixes compatibility with Web Modeler OpenAPI-generated templates that use Unix epoch timestamps

Agent-Logs-Url: https://github.com/camunda/connectors/sessions/514b81bf-c6eb-4434-ac55-fa1c98814811

+++ element-template-generator/core/src/main/java/io/camunda/connector/generator/dsl/CommonProperties.java
@@ -70,7 +70,7 @@ public class CommonProperties {
     return builder;
   }
 
-  public static PropertyBuilder version(Integer version) {
+  public static PropertyBuilder version(Long version) {
     return HiddenProperty.builder()
         .id("version")
         .group("connector")
